Extract layer integrations
