+++ lib/assets/javascripts/cartodb3/data/layer-definitions-collection.js
@@ -5,15 +5,6 @@ var layerLetters = require('./layer-letters');
 var nodeIds = require('./analysis-definition-node-ids.js');
 var camshaftReference = require('./camshaft-reference');
 
-var LAYER_TYPE_TO_LAYER_CREATE_METHOD = {
-  'cartodb': 'createCartoDBLayer',
-  'gmapsbase': 'createGMapsBaseLayer',
-  'plain': 'createPlainLayer',
-  'tiled': 'createTileLayer',
-  'torque': 'createTorqueLayer',
-  'wms': 'createWMSLayer'
-};
-
 /**
  * Collection of layer definitions
  */
@@ -82,22 +73,17 @@ module.exports = Backbone.Collection.extend({
 
   initialize: function (models, options) {
     if (!options.configModel) throw new Error('configModel is required');
-    if (!options.visMap) throw new Error('visMap (vis.map) is required');
     if (!options.analysisDefinitionsCollection) throw new Error('analysisDefinitionsCollection is required');
     if (!options.analysisDefinitionNodesCollection) throw new Error('analysisDefinitionNodesCollection is required');
     if (!options.mapId) throw new Error('mapId is required');
 
     this._configModel = options.configModel;
-    this._visMap = options.visMap;
     this._analysisDefinitionsCollection = options.analysisDefinitionsCollection;
     this._analysisDefinitionNodesCollection = options.analysisDefinitionNodesCollection;
 
     this.mapId = options.mapId;
 
-    this.on('add', this._onAdd, this);
     this.on('sync', this._onSync, this);
-    this.on('change', this._onChange, this);
-    this.on('remove', this._onRemove, this);
   },
 
   url: function () {
@@ -132,18 +118,7 @@ module.exports = Backbone.Collection.extend({
     return this._analysisDefinitionNodesCollection;
   },
 
-  _onAdd: function (m) {
-    // If added but not yet saved, postpone the creation until persisted (see sync listener)
-    if (!m.isNew()) {
-      this._createLayer(m);
-    }
-  },
-
   _onSync: function (m) {
-    if (!this._getLayer(m)) {
-      this._createLayer(m);
-    }
-
     // Create analysis definition if there is none yet
     var nodeId = m.get('source');
     var analysisDefinitionModel = this._analysisDefinitionsCollection.findByNodeId(nodeId);
@@ -157,43 +132,6 @@ module.exports = Backbone.Collection.extend({
     this._analysisDefinitionsCollection.create({analysis_definition: nodeModel.toJSON()});
   },
 
-  _onChange: function (m) {
-    var attrs = m.changedAttributes();
-    var layer = this._getLayer(m);
-
-    if (!m.isNew()) {
-      if (attrs.type) {
-        layer.remove();
-        this._createLayer(m);
-      } else {
-        layer.update(attrs);
-      }
-    }
-  },
-
-  _onRemove: function (m) {
-    if (!m.isNew()) {
-      this._getLayer(m).remove();
-    }
-  },
-
-  _getLayer: function (m) {
-    return this._visMap.getLayerById(m.id);
-  },
-
-  _createLayer: function (m) {
-    var attrs = this._deepClone(m.attributes);
-    delete attrs.source; // a new layer should not have an analyse to begin with, even if we have it in the context of editor
-
-    var createMethodName = LAYER_TYPE_TO_LAYER_CREATE_METHOD[attrs.type.toLowerCase()];
-    if (!createMethodName) throw new Error('no create method name found for type ' + attrs.type);
-
-    var createMethod = this._visMap[createMethodName];
-    if (!_.isFunction(createMethod)) throw new Error(createMethodName + ' is not a function');
-
-    createMethod.call(this._visMap, attrs);
-  },
-
   _letters: function (otherLetters) {
     var lettersFromAddedModels = _.compact(this.pluck('letter'));
 
@@ -202,9 +140,5 @@ module.exports = Backbone.Collection.extend({
     var lettersFromModelsNotYetAdded = _.chain(this._byId).values().invoke('get', 'letter').value();
 
     return _.union(lettersFromAddedModels, lettersFromModelsNotYetAdded, otherLetters);
-  },
-
-  _deepClone: function (obj) {
-    return JSON.parse(JSON.stringify(obj)); // deep clone;
   }
 });
