+++ lib/assets/javascripts/cartodb3/deep-insights-integrations.js
@@ -3,9 +3,15 @@
  */
 var F = function (opts) {
   if (!opts.deepInsightsDashboard) throw new Error('deepInsightsDashboard is required');
+  if (!opts.layerDefinitionsCollection) throw new Error('layerDefinitionsCollection is required');
   if (!opts.widgetDefinitionsCollection) throw new Error('widgetDefinitionsCollection is required');
 
   this._diDashboard = opts.deepInsightsDashboard;
+
+  opts.layerDefinitionsCollection.on('add', this._onLayerDefinitionAdded, this);
+  opts.layerDefinitionsCollection.on('sync', this._onLayerDefinitionSynced, this);
+  opts.layerDefinitionsCollection.on('change', this._onLayerDefinitionChanged, this);
+  opts.layerDefinitionsCollection.on('remove', this._onLayerDefinitionRemoved, this);
 
   opts.widgetDefinitionsCollection.on('add', this._onWidgetDefinitionAdded, this);
   opts.widgetDefinitionsCollection.on('sync', this._onWidgetDefinitionSynced, this);
@@ -13,7 +19,7 @@ var F = function (opts) {
   opts.widgetDefinitionsCollection.on('destroy', this._onWidgetDefinitionDestroyed, this);
 };
 
-F.prototype._visMap = function () {
+F.prototype.visMap = function () {
   return this._diDashboard.getMap().map;
 };
 
@@ -63,12 +69,68 @@ F.prototype._createWidgetModel = function (m) {
   var methodName = 'create' + infix + 'Widget';
 
   var layerId = m.get('layer_id');
-  var layerModel = this._visMap().getLayerById(layerId);
+  var layerModel = this.visMap().getLayerById(layerId);
 
   var widgetModel = this._diDashboard[methodName](m.attributes, layerModel);
   if (widgetModel) {
     widgetModel.set('show_stats', true);
   }
+};
+
+F.prototype._onLayerDefinitionAdded = function (m) {
+  // If added but not yet saved, postpone the creation until persisted (see sync listener)
+  if (!m.isNew() && !this._getLayer(m)) {
+    this._createLayer(m);
+  }
+};
+
+F.prototype._onLayerDefinitionSynced = function (m) {
+  if (!this._getLayer(m)) {
+    this._createLayer(m);
+  }
+};
+
+F.prototype._onLayerDefinitionChanged = function (m) {
+  var attrs = m.changedAttributes();
+  var layer = this._getLayer(m);
+
+  if (!m.isNew()) {
+    if (attrs.type) {
+      layer.remove();
+      this._createLayer(m);
+    } else {
+      layer.update(attrs);
+    }
+  }
+};
+
+F.prototype._onLayerDefinitionRemoved = function (m) {
+  if (!m.isNew()) {
+    this._getLayer(m).remove();
+  }
+};
+
+var LAYER_TYPE_TO_LAYER_CREATE_METHOD;
+F.prototype._createLayer = function (m) {
+  var attrs = JSON.parse(JSON.stringify(m.attributes)); // deep clone;
+  delete attrs.source; // a new layer should not have an analyse to begin with, even if we have it in the context of editor
+
+  LAYER_TYPE_TO_LAYER_CREATE_METHOD = LAYER_TYPE_TO_LAYER_CREATE_METHOD || {
+    'cartodb': 'createCartoDBLayer',
+    'gmapsbase': 'createGMapsBaseLayer',
+    'plain': 'createPlainLayer',
+    'tiled': 'createTileLayer',
+    'torque': 'createTorqueLayer',
+    'wms': 'createWMSLayer'
+  };
+  var createMethodName = LAYER_TYPE_TO_LAYER_CREATE_METHOD[attrs.type.toLowerCase()];
+  if (!createMethodName) throw new Error('no create method name found for type ' + attrs.type);
+
+  this.visMap()[createMethodName](attrs);
+};
+
+F.prototype._getLayer = function (m) {
+  return this.visMap().getLayerById(m.id);
 };
 
 module.exports = F;
