+++ lib/assets/javascripts/cartodb3/editor.js
@@ -101,7 +101,6 @@ deepInsights.createDashboard('#dashboard', vizJSON, {
 
   var layerDefinitionsCollection = new LayerDefinitionsCollection([], {
     configModel: configModel,
-    visMap: vis.map,
     analysisDefinitionsCollection: analysisDefinitionsCollection,
     analysisDefinitionNodesCollection: analysisDefinitionNodesCollection,
     mapId: mapId
@@ -118,6 +117,7 @@ deepInsights.createDashboard('#dashboard', vizJSON, {
   });
   var deepInsightsIntegrations = new DeepInsightsIntegrations({
     deepInsightsDashboard: dashboard,
+    layerDefinitionsCollection: layerDefinitionsCollection,
     widgetDefinitionsCollection: widgetDefinitionsCollection
   });
