+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/add-widgets-view.spec.js
@@ -23,7 +23,6 @@ describe('components/modals/add-widgets/add-widgets-view', function () {
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: this.configModel,
-      visMap: this.vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: analysisDefinitionNodesCollection,
       mapId: 'map-123'
@@ -116,7 +115,6 @@ describe('components/modals/add-widgets/add-widgets-view', function () {
 
       var layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
         configModel: this.configModel,
-        visMap: this.vis.map,
         analysisDefinitionsCollection: {},
         analysisDefinitionNodesCollection: analysisDefinitionNodesCollection,
         mapId: 'map-123'
