+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/create-tuples-items.spec.js
@@ -4,7 +4,6 @@ var AnalysisDefinitionNodesCollection = require('../../../../../../javascripts/c
 var AnalysisTableModel = require('../../../../../../javascripts/cartodb3/data/analysis-table-model');
 var LayerDefinitionsCollection = require('../../../../../../javascripts/cartodb3/data/layer-definitions-collection');
 var createTuplesItems = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/create-tuples-items');
-var createDefaultVis = require('../../../create-default-vis');
 
 describe('components/modals/add-widgets/create-tuples-items', function () {
   beforeEach(function () {
@@ -48,7 +47,6 @@ describe('components/modals/add-widgets/create-tuples-items', function () {
       configModel: configModel
     });
 
-    var vis = createDefaultVis();
     this.layerDefinitionsCollection = new LayerDefinitionsCollection([
       {
         id: 'l0',
@@ -70,7 +68,6 @@ describe('components/modals/add-widgets/create-tuples-items', function () {
       }
     ], {
       configModel: configModel,
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: analysisDefinitionNodesCollection,
       mapId: 'm-123'
