+++ lib/assets/test/spec/cartodb3/data/background-importer/geocoding-collection.spec.js
@@ -20,7 +20,6 @@ describe('common/background-polling/geocodings-collection', function () {
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection([], {
       configModel: this.configModel,
-      visMap: this.vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: {},
       mapId: 123
