+++ lib/assets/test/spec/cartodb3/data/layer-definitions-collection.spec.js
@@ -26,7 +26,6 @@ describe('data/layer-definitions-collection', function () {
 
     this.collection = new LayerDefinitionsCollection(null, {
       configModel: configModel,
-      visMap: this.vis.map,
       analysisDefinitionsCollection: this.analysisDefinitionsCollection,
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
       mapId: 'm-123'
@@ -168,69 +167,6 @@ describe('data/layer-definitions-collection', function () {
 
       expect(this.l2.get('source')).toEqual('b0');
       expect(this.analysisDefinitionNodesCollection.get('b0')).toBeDefined();
-    });
-  });
-
-  describe('when a new layer is created', function () {
-    beforeEach(function () {
-      this.collection.create({
-        id: 'integration-test',
-        options: {
-          type: 'Plain',
-          color: 'blue'
-        }
-      });
-      this.layerDefinitionModel = this.collection.get('integration-test');
-    });
-
-    it('should have created the layer', function () {
-      var l = this.vis.map.layers.get(this.layerDefinitionModel.id);
-      expect(l).toBeDefined();
-      expect(l.get('color')).toEqual('blue');
-      expect(l.get('type')).toEqual('Plain');
-    });
-
-    describe('when update some layer attrs', function () {
-      beforeEach(function () {
-        this.layerDefinitionModel.set({
-          color: 'pink',
-          letter: 'c'
-        });
-      });
-
-      it('should update the equivalent layer', function () {
-        var l = this.vis.map.layers.get(this.layerDefinitionModel.id);
-        expect(l.get('color')).toEqual('pink');
-      });
-    });
-
-    describe('when update layer includes change of type', function () {
-      beforeEach(function () {
-        this.layerBefore = this.vis.map.layers.get(this.layerDefinitionModel.id);
-        this.layerDefinitionModel.set({
-          type: 'CartoDB',
-          table_name: 'my_table',
-          cartocss: 'asd',
-          sql: 'SELECT * FROM my_table'
-        });
-        this.layerAfter = this.vis.map.layers.get(this.layerDefinitionModel.id);
-      });
-
-      it('should have re-created layer', function () {
-        expect(this.layerAfter).not.toBe(this.layerBefore);
-        expect(this.layerAfter.get('sql')).toEqual('SELECT * FROM my_table');
-        expect(this.layerAfter.get('type')).toEqual('CartoDB');
-      });
-    });
-
-    describe('when removing layer', function () {
-      beforeEach(function () {
-        this.layerDefinitionModel.destroy();
-      });
-
-      it('should no longer be accessible', function () {
-        expect(this.vis.map.layers.get(this.layerDefinitionModel.id)).toBeUndefined();
-      });
     });
   });
 });
