+++ lib/assets/test/spec/cartodb3/deep-insights-integrations.spec.js
@@ -26,11 +26,13 @@ describe('deep-insights-integrations', function () {
         throw new Error('error creating dashboard ' + error);
       }
       this.dashboard = dashboard;
-      this.vis = dashboard.getMap();
+
+      this.layerDefinitionsCollection = new Backbone.Collection();
       this.widgetDefinitionsCollection = new Backbone.Collection();
 
       this.integrations = new DeepInsightsIntegrations({
         deepInsightsDashboard: dashboard,
+        layerDefinitionsCollection: this.layerDefinitionsCollection,
         widgetDefinitionsCollection: this.widgetDefinitionsCollection
       });
 
@@ -66,7 +68,7 @@ describe('deep-insights-integrations', function () {
         column: 'col',
         operation: 'avg'
       }));
-      expect(args[1]).toBe(this.vis.map.layers.first());
+      expect(args[1]).toBe(this.integrations.visMap().layers.first());
     });
 
     it('should enable show_stats for the created widget model', function () {
@@ -124,6 +126,66 @@ describe('deep-insights-integrations', function () {
 
       it('should remove the corresponding widget model', function () {
         expect(this.widgetModel.remove).toHaveBeenCalled();
+      });
+    });
+  });
+
+  describe('when a new layer is created', function () {
+    beforeEach(function () {
+      this.layerDefinitionModel = this.layerDefinitionsCollection.add({
+        id: 'integration-test',
+        type: 'Plain',
+        color: 'blue'
+      });
+    });
+
+    it('should have created the layer', function () {
+      var l = this.integrations.visMap().layers.get(this.layerDefinitionModel.id);
+      expect(l).toBeDefined();
+      expect(l.get('color')).toEqual('blue');
+      expect(l.get('type')).toEqual('Plain');
+    });
+
+    describe('when update some layer attrs', function () {
+      beforeEach(function () {
+        this.layerDefinitionModel.set({
+          color: 'pink',
+          letter: 'c'
+        });
+      });
+
+      it('should update the equivalent layer', function () {
+        var l = this.integrations.visMap().layers.get(this.layerDefinitionModel.id);
+        expect(l.get('color')).toEqual('pink');
+      });
+    });
+
+    describe('when update layer includes change of type', function () {
+      beforeEach(function () {
+        this.layerBefore = this.integrations.visMap().layers.get(this.layerDefinitionModel.id);
+        this.layerDefinitionModel.set({
+          type: 'CartoDB',
+          table_name: 'my_table',
+          cartocss: 'asd',
+          sql: 'SELECT * FROM my_table'
+        });
+        this.layerAfter = this.integrations.visMap().layers.get(this.layerDefinitionModel.id);
+      });
+
+      it('should have re-created layer', function () {
+        expect(this.layerAfter).not.toBe(this.layerBefore);
+        expect(this.layerAfter.get('sql')).toEqual('SELECT * FROM my_table');
+        expect(this.layerAfter.get('type')).toEqual('CartoDB');
+      });
+    });
+
+    describe('when removing layer', function () {
+      beforeEach(function () {
+        this.layerDefinitionsCollection.remove(this.layerDefinitionModel);
+      });
+
+      it('should no longer be accessible', function () {
+        expect(this.integrations.visMap().layers.get(this.layerDefinitionModel.id)).toBeUndefined();
       });
     });
   });
