+++ lib/assets/test/spec/cartodb3/editor/layers/analysis-view/composite-layers-analysis-view.spec.js
@@ -40,7 +40,6 @@ describe('editor/layers/analysis-views/composite-layer-analysis-view', function 
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: {},
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
       mapId: 'map-123'
