+++ lib/assets/test/spec/cartodb3/editor/layers/layer-analysis-views.spec.js
@@ -15,7 +15,6 @@ describe('editor/layers/layer-analyses-view', function () {
     var vis = createDefaultVis();
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: {},
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
       mapId: 'm-123'
