+++ lib/assets/test/spec/cartodb3/editor/layers/layer-content-view/analyses-form-types/analysis-source-options-model.spec.js
@@ -24,7 +24,6 @@ describe('editor/layers/layer-content-views/analyses-form-types/analysis-source-
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: configModel,
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
       mapId: 'map-123'
