+++ lib/assets/test/spec/cartodb3/editor/layers/layer-content-view/analyses-view.spec.js
@@ -43,7 +43,6 @@ describe('editor/layers/layer-content-view/analyses-view', function () {
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: configModel,
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: analysisDefinitionNodesCollection,
       mapId: 'map-123'
