+++ lib/assets/test/spec/cartodb3/editor/layers/layer-content-view/analysis-form-type-view.spec.js
@@ -45,7 +45,6 @@ describe('editor/layers/layer-content-view/analyses-form-type-view', function ()
 
     this.layerDefinitionsCollection = new LayerDefinitionsCollection(null, {
       configModel: configModel,
-      visMap: vis.map,
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
       mapId: 'map-123'
