+++ lib/assets/test/spec/cartodb3/editor/layers/layers-view.spec.js
@@ -29,8 +29,7 @@ describe('editor/layers/layers-view', function () {
       configModel: {},
       analysisDefinitionsCollection: {},
       analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
-      mapId: 'm-123',
-      visMap: vis.map
+      mapId: 'm-123'
     });
     this.layerDefinitionsCollection.add({
       id: 'l1',
